+++ HoursBank/HoursBank.Service/Services/UserService.cs
@@ -27,9 +27,20 @@ namespace HoursBank.Service.Services
             var result = await _repository.SelectAsync(id);
             if (result != null)
             {
+                result.Password = BHCrypto.Decode(result.Password);
                 return _mapper.Map<UserResponse>(result);
             }
             return null;
+        }
+
+        public async Task<bool> IsAdmin(int id)
+        {
+            var result = await _repository.SelectAsync(id);
+            if (result != null)
+            {
+                return result.Admin;
+            }
+            return false;
         }
 
         public async Task<IEnumerable<UserResponse>> GetAll()
@@ -37,7 +48,21 @@ namespace HoursBank.Service.Services
             var result = await _repository.SelectAsync();
             if (result.Any())
             {
-                return _mapper.Map<List<UserResponse>>(result);
+                var listResult = _mapper.Map<List<UserResponse>>(result);
+                listResult.ForEach(r => { r.Password = ""; });
+                return listResult;
+            }
+            return null;
+        }
+
+        public async Task<IEnumerable<UserResponse>> GetUsersToApprove()
+        {
+            var result = await _repository.SelectAsync();
+            if (result.Any())
+            {
+                var listResult = _mapper.Map<List<UserResponse>>(result.Where(u => !u.Active.HasValue));
+                listResult.ForEach(r => { r.Password = ""; });
+                return listResult;
             }
             return null;
         }
@@ -48,6 +73,7 @@ namespace HoursBank.Service.Services
             if (users.Any())
             {
                 var result = users.FirstOrDefault(c => c.Email == email);
+                result.Password = BHCrypto.Decode(result.Password);
                 return _mapper.Map<UserResponse>(result);
             }
             return null;
@@ -59,6 +85,10 @@ namespace HoursBank.Service.Services
             if (users.Any())
             {
                 var result = users.Where(c => c.TeamId == id).ToList();
+                result.ForEach(r =>
+                {
+                    r.Password = BHCrypto.Decode(r.Password);
+                });
                 return _mapper.Map<List<UserResponse>>(result);
             }
             return null;
@@ -71,6 +101,7 @@ namespace HoursBank.Service.Services
             var result = await _repository.InsertAsync(entity);
             if (result != null)
             {
+                result.Password = BHCrypto.Decode(result.Password);
                 return _mapper.Map<UserResponse>(result);
             }
             return null;
@@ -78,23 +109,47 @@ namespace HoursBank.Service.Services
 
         public async Task<UserResponse> Put(UserDto user)
         {
+            if (user.Id != 1)
+            {
+                var userOld = await _repository.SelectAsync(user.Id);
+
+                if (!string.IsNullOrEmpty(user.Password) && !user.Password.Equals(BHCrypto.Decode(userOld.Password)))
+                {
+                    userOld.Password = BHCrypto.Encode(user.Password);
+                }
+
+                userOld.Name = user.Name;
+                userOld.ClientId = user.ClientId;
+                userOld.ClientSecret = user.ClientSecret;
+                userOld.Email = user.Email;
+                userOld.TeamId = user.TeamId.HasValue ? user.TeamId : userOld.TeamId;
+                userOld.Active = user.Active;
+
+                var result = await _repository.UpdateAsync(userOld);
+                if (result != null)
+                {
+                    result.Password = BHCrypto.Decode(result.Password);
+                    return _mapper.Map<UserResponse>(result);
+                }
+            }
+            return null;
+        }
+
+        public async Task<UserResponse> Approve(UserDto user)
+        {
             var userOld = await _repository.SelectAsync(user.Id);
 
-            if(!string.IsNullOrEmpty(user.Password) && !user.Password.Equals(BHCrypto.Decode(userOld.Password)))
+            userOld.Active = user.Active;
+            if (user.Active.HasValue && user.Active.Value)
             {
-                userOld.Password = BHCrypto.Encode(user.Password);
+                userOld.TeamId = user.TeamId.Value;
             }
 
-            userOld.Name = user.Name;
-            userOld.Active = user.Active;
-            userOld.ClientId = user.ClientId;
-            userOld.ClientSecret = user.ClientSecret;
-            userOld.Email = user.Email;
-            userOld.TeamId = user.TeamId.HasValue ? user.TeamId : userOld.TeamId;
 
             var result = await _repository.UpdateAsync(userOld);
             if (result != null)
             {
+                result.Password = BHCrypto.Decode(result.Password);
                 return _mapper.Map<UserResponse>(result);
             }
             return null;
@@ -102,7 +157,14 @@ namespace HoursBank.Service.Services
 
         public async Task<bool> Delete(int id)
         {
-            return await _repository.DeleteAsync(id);
+            if (id != 1)
+            {
+                var user = await this.Get(id);
+                if (user != null && (!user.Active.HasValue || !user.Active.Value)) {
+                    return await _repository.DeleteAsync(id);
+                }
+            }
+            return false;
         }
     }
 }
